+++ flowmachine/flowmachine/features/subscriber/proportion_outgoing.py
@@ -16,7 +16,7 @@ from ..utilities import EventsTablesUnion
 from .metaclasses import SubscriberFeature
 
 
-class ProportionEventOutgoing(SubscriberFeature):
+class ProportionOutgoing(SubscriberFeature):
     """
     Find the proportion of interactions initiated by a
     given subscriber.
@@ -47,7 +47,7 @@ class ProportionEventOutgoing(SubscriberFeature):
 
     Examples
     --------
-    >>> ProportionEventOutgoing("2016-01-01", "2016-01-02",
+    >>> ProportionOutgoing("2016-01-01", "2016-01-02",
     table="events.calls")
                    msisdn  proportion_outgoing  proportion_incoming
     0    jWlyLwbGdvKV35Mm                    1                    0