+++ flowmachine/tests/test_proportion_outgoing.py
@@ -8,19 +8,19 @@ Tests for the feature subscriber event proportions.
 
 import pytest
 
-from flowmachine.features.subscriber.proportion_outgoing import ProportionEventOutgoing
+from flowmachine.features.subscriber.proportion_outgoing import ProportionOutgoing
 
 
 """
-Tests for ProportionEventOutgoing() feature class.
+Tests for ProportionOutgoing() feature class.
 """
 
 
 def test_returns_correct_column_names():
     """
-    ProportionEventOutgoing() dataframe has expected column names.
+    ProportionOutgoing() dataframe has expected column names.
     """
-    ud = ProportionEventOutgoing("2016-01-01", "2016-01-04")
+    ud = ProportionOutgoing("2016-01-01", "2016-01-04")
     assert [
         "subscriber",
         "proportion_outgoing",
@@ -30,9 +30,9 @@ def test_returns_correct_column_names():
 
 def test_returns_correct_values(get_dataframe):
     """
-    ProportionEventOutgoing() dataframe contains expected values.
+    ProportionOutgoing() dataframe contains expected values.
     """
-    ud = ProportionEventOutgoing("2016-01-01", "2016-01-04")
+    ud = ProportionOutgoing("2016-01-01", "2016-01-04")
     df1 = get_dataframe(ud).set_index("subscriber")
     assert 0.600000 == df1.loc["ZM3zYAPqx95Rw15J"]["proportion_outgoing"]
     assert 0.400000 == df1.loc["ZM3zYAPqx95Rw15J"]["proportion_incoming"]
@@ -40,7 +40,7 @@ def test_returns_correct_values(get_dataframe):
 
 def test_passing_not_known_subscriber_identifier_raises_error():
     """
-    ProportionEventOutgoing() passing not know `subscriber_identifier` raises error.
+    ProportionOutgoing() passing not know `subscriber_identifier` raises error.
     """
     with pytest.raises(ValueError):
-        ProportionEventOutgoing("2016-01-01", "2016-01-04", subscriber_identifier="foo")
+        ProportionOutgoing("2016-01-01", "2016-01-04", subscriber_identifier="foo")